+++ test/matchers/document_test.rb
@@ -13,5 +13,7 @@ describe "Document" do
 
     it { must have_field(:name).of_type(String).with_default_value("me") }
     it { must have_fields(:name, :login).of_type(String).with_default_value("me") }
+
+    it { must have_index_for(:name) }
   end
 end